communityPulse-nova-rangers/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
firebase-admin>=6.6.0
google-cloud-firestore>=2.19.0
google-cloud-pubsub>=2.27.0
google-cloud-bigquery>=3.27.0
google-adk>=0.3.0
sentence-transformers>=3.3.0
chromadb>=0.5.23
pytesseract>=0.3.13
opencv-python-headless>=4.10.0
pdfplumber>=0.11.4
geopy>=2.4.1
python-geohash>=0.8.5
apscheduler>=3.10.4
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-gcp-trace>=1.8.0
upstash-redis>=1.1.0
resend>=2.5.0
python-multipart>=0.0.18
httpx>=0.28.0
fastapi-cloudauth>=0.4.3
python-dotenv>=1.0.1
google-generativeai>=0.8.0
google-generativeai
